refactor: modified directory name
